requirements.txt
click==7.1.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
imagezmq==1.1.1
imutils==0.5.4
netifaces==0.10.9
numpy==1.19.5
opencv-python==4.5.1.48
paho-mqtt==1.5.1
Pillow==8.1.0
pyzmq==21.0.1
